Update bot.py
Updated the /alltime command to start on page zero instead of page one but to omit the `anonymous` user. This was done in case another user was to surpass the `anonymous` user on the rankings.

Updated the /monthly command to start on page zero instead of page one but to omit the `anonymous` user. This was done in case another user was to surpass the `anonymous` user on the rankings.